requirements.txt
asgiref==3.7.2
asttokens==2.3.0
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.2.0
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
Django==4.2.5
django-crispy-forms==2.0
django-filter==23.3
djangorestframework==3.14.0
exceptiongroup==1.1.3
executing==1.2.0
fonttools==4.42.1
idna==3.4
iniconfig==2.0.0
ipykernel==6.25.1
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
numpy==1.25.2
packaging==23.1
pandas==2.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
pluggy==1.3.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pyparsing==3.0.9
pytest==7.4.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.1
requests==2.31.0
scikit-learn==1.3.1
scipy==1.11.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
threadpoolctl==3.2.0
tomli==2.0.1
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
wcwidth==0.2.6
